Migrate vault.yml from ansible-vault to plain yaml
The file was already encrypted with git-crypt, so also encrypting it
with ansible-vault was unnecessary and just adding more overhead. This
is here so that we can reference these values from the defaults file,
without having to encrypt the defaults file. This way it's discoverable
by others who might want to use this someday.