Add monitoring, mbsync plus others, and some refactoring
I finally reconciled a lot of changes I'd been experimenting with
locally, but hadn't committed obviously. So now those are tracked
properly, as well as Moinmoin, Mbsync, and others. Also refactored
heavily the docker-compose modules tasks file, to use with_items with
lists of files, templates, and directories, instead of having endless
blocks of the same types. Seems a lot cleaner and easier to read, time
will tell if it's a better approach and something that will last.