Migrate docker-compose file to separate files and misc fixes
The old docker-compose.yml file was getting a bit unmanageable, and so I
decided to split it out into a number of smaller files. They are still
all sharing the common .env file for secrets and things currently, so
this was a minor change to allow easier management of things for now. I
also update some of the images and the prometheus config accordingly.