Refactor authorized_keys task to use jinja filter
The authorized_key module is built in such a way that makes it almost
useless when used with lists and iterators. It can't have ownership over
the authorized_keys file, and hence can't remove keys, that shouldn't be
present. This is a huge oversight, and something that should be fixed in
the module, so that this is the default behaviour.

In the meantime, I was able to use a Jinja filter, to join a list of
keys, using newlines, and send them to a single invocation of the
authorized_keys module. Thus getting around the issue, and allowing the
module to have exclusive access to the file, and purge old entries that
shouldn't be present any longer.